// src/app/Enums/hobbies.ts
export enum Hobbies
{
    None = 0,
    Reading = 1,
    Writing = 2,
    Painting = 3,
    Cooking = 4,
    Gardening = 5,
    Hiking = 6,
    Swimming = 7,
    Photography = 8,
    PlayingMusic = 9,
    Dancing = 10,
    Fishing = 11,
    Cycling = 12,
    Traveling = 13,
    BirdWatching = 14,
    Gaming = 15,
    Yoga = 16,
    Chess = 17,
    Sculpting = 18,
    Woodworking = 19,
    Knitting = 20,
    CollectingCoins = 21,
    Astronomy = 22,
    Surfing = 23,
    Skiing = 24,
    Origami = 25,
    Baking = 26,
    Pottery = 27,
    Billiards = 28,
    Camping = 29,
    AutoRepair = 30,
    Calligraphy = 31,
    ModelBuilding = 32,
    BirdKeeping = 33,
    Antiquing = 34,
    Volunteering = 35,
    Running = 36,
    TableTennis = 37,
    ScubaDiving = 38,
    Meditation = 39,
    HorsebackRiding = 40,
    Archery = 41,
    Juggling = 42,
    WineTasting = 43,
    KiteFlying = 44,
    Rowing = 45,
    Stargazing = 46,
    Beekeeping = 47,
    CrosswordPuzzles = 48,
    StandUpComedy = 49,
    ArmWrestling = 50,
}